NetLogo Style Moves
Full migration to Netlogo style movement.

+++ src/@organism/organism.m
@@ -6,7 +6,7 @@ classdef (Abstract) organism < handle
         Species string % Name of the species, e.g. sheep, kangaroo, dingo
         FeedsOn string % Name of the species that we feed on
         Colour (3, 1) double % Default colour to plot this cell as
-        Coordinate (2, 1) int16 % Coordinate that this organism is at
+        Coordinate (2, 1) double % Coordinate that this organism is at
         ProbReproduce (1,1) double % Probability of breeding this turn
         Energy (1,1) double % Energy when zero organism dies
         GainFromFood (1,1) double % How much energy gain from one food
